Continue to webpackize the data manager app
 - es6 imports for DAO, luxon, axios

Still a few things in the global namespace
