+++ modules/admin/app/assets/js/datamanager/components/_comp.vue
