+++ modules/admin/app/assets/js/datamanager/components/delete-button.vue
