+++ modules/admin/app/assets/js/datamanager/components/drag-handle.vue
