+++ modules/admin/app/assets/js/datamanager/components/file-picker.vue
