+++ modules/admin/app/assets/js/datamanager/components/log-window.vue
