+++ modules/admin/app/assets/js/datamanager/components/mixin-preview-panel.vue
