+++ modules/admin/app/assets/js/datamanager/components/mixin-stage.vue
