+++ modules/admin/app/assets/js/datamanager/components/mixin-two-panel.vue
