+++ modules/admin/app/assets/js/datamanager/components/modal-window.vue
