+++ modules/admin/app/assets/js/datamanager/components/panel-convert-preview.vue
