+++ modules/admin/app/assets/js/datamanager/components/panel-file-preview.vue
