+++ modules/admin/app/assets/js/datamanager/components/transformation-editor.vue
