+++ modules/admin/app/assets/js/datamanager/components/transformation-item.vue
