+++ modules/admin/app/assets/js/datamanager/components/upload-progress.vue
