+++ modules/admin/app/assets/js/datamanager/components/xquery-editor.vue
