+++ modules/admin/app/assets/js/datamanager/components/xslt-editor.vue
