+++ modules/admin/app/assets/js/datamanager/worker.js
