+++ webpack.config.js
@@ -29,7 +29,7 @@ module.exports = {
         }
       }
     ]
-  },
+    },
   resolve: {
     extensions: ['.ts', '.js', '.vue']
   }
